# Bug #87483 Accelerated AES with ARMv8 Crypto Extensions in yaSSL