Support for 64-bit ARMv8 AES cryptography extensions in yaSSL taoCrypt.
The AES with ARMv8 NEON intrinsics will optimized the performance rather than uses table-based lookup.

Environment: Hislicon Taishan D03/D05 , Cavium ThunderX , AMD seattle (Softiron)	
OS: ubuntu16.04/Debian 8 ,  Centos

use yassl taocrypt benchmark(SOURCE_HOME/extra/yassl/taocrypt/benchmark/benchmark.cpp):

Platform \ Case         	       Software AES	AArch64 AES Intrinsics
AMD seattle (Softiron)	        92.96 MB/s	846.04 MB/s
Hisilicon Taishan(D03/A57)	99.01 MB/s	887.31 MB/s
Hisilicon Taishan(D05/A72)	113.60 MB/s	1218.64 MB/s
Cavium ThunderX	                47.36 MB/s	151.09 MB/s