@guyuqi 2017-08-22
I confirm the code being submitted is offered under the terms of the OCA, and that I am authorized to contribute it.